requirements.txt
streamlit
openai
langchain
langchain-openai
langchain-google-genai
langchain-community
langgraph
tmdbv3api
tavily-python
python-dotenv
faker
pandas
numpy
requests
PyPDF2
pdfplumber
scikit-learn
plotly
